# perf(imports): Bulk import CSV transactions