--- a/Gemfile
+++ b/Gemfile
@@ -58,6 +58,7 @@ gem "intercom-rails"
 gem "plaid"
 gem "rotp", "~> 6.3"
 gem "rqrcode", "~> 2.2"
+gem "activerecord-import"
 
 group :development, :test do
   gem "debug", platforms: %i[mri windows]