Bulk import transactions from CSV instead of importing each row at a time resulting in a optimization in performance and faster imports

| Entries | Before: (s)  | After: (s) |
|:---:|:---:|:---:|
| 1000 | 5.7 | 2.2 |
| 10000 |    93.9   | 16.4 |
| 100000 | 2083.7 | 203.8|

Fixes: #1846.